+++ example/lib/screens/home_page.dart
@@ -25,9 +25,9 @@ class HomePage extends StatelessWidget {
                     () => QR.to(
                         "/params?test=${Random().nextInt(1000)}&go=${Random().nextInt(1000)}")),
                 QButton("Test not found Page", () => QR.to("/parent/no-child")),
-                QButton("Site Map", () => QR.to("/sitemap")),
-                QButton("Test Nested Navigation", () => QR.to("/nested")),
-                QButton("Go to Login/Home-Setting Case", () => QR.to('/login')),
+                QButton("Add Remove Routes", () => QR.to("/add-remove-routes")),
+                QButton("Nested Navigation", () => QR.to("/nested")),
+                QButton("Overlays", () => QR.to("/overlays")),
               ],
             ))));
   }