+++ lib/qlevar_router.dart
@@ -1,5 +1,6 @@
 import 'src/qr.dart';
 
+export 'src/overlays/qdialog.dart';
 export 'src/pages/qpages.dart';
 export 'src/routers/qrouter.dart';
 export 'src/routers/router_delegate.dart';