+++ lib/src/routes/qroute_internal.dart
@@ -44,7 +44,10 @@ class QRouteInternal {
 
   factory QRouteInternal.from(QRoute route, String cureentPath) {
     final key = QKey(route.name ?? route.path);
-    final fullPath = '${cureentPath == '' ? '' : '$cureentPath'}${route.path}';
+    if (!route.path.startsWith('/')) {
+      route = route.copyWith(path: '/${route.path}');
+    }
+    final fullPath = '$cureentPath${route.path}';
     QR.treeInfo.namePath[route.name ?? route.path] = fullPath;
     return QRouteInternal(
         key: key,